Completing testing and creating Readme file.

+++ calculator/src/main/java/questions/CalculatorOperations.java
@@ -2,11 +2,14 @@ package questions;
 
 import java.util.function.BinaryOperator;
 
+/*
+ * Here we can add our own operations.
+ */
 public class CalculatorOperations implements CommonOperations {
 	
 	// Subtraction
-	final static BinaryOperator<Integer> subtraction = (a, b) -> a-b;
+	final static BinaryOperator<Double> subtraction = (a, b) -> a-b;
 	
 	// Division
-	final static BinaryOperator<Integer> division = (a, b) -> a/b;
+	final static BinaryOperator<Double> division = (a, b) -> a/b;
 }
